# [Traceql Metrics] Variety of performance improvements